Create AKORN.yaml
Dear Tonkeeper Team,

This token is a subtoken of the AKA token. In the AKA mini-app on Telegram, this token is scheduled to be distributed as part of the rewards. A portion of the rewards also consists of the AKA token itself. Additionally, the name of this token is mentioned on the AKA website.

If you need any further information, please let us know.

Kind regards,
Aka Team